+++ PluginsAndFeatures/azure-toolkit-for-rider/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## [Unreleased]
 
+### Changed
+- Update tool window icon
+- Reimplement Azurite support
+
 ## [4.0.0-preview.1] - 2024-01-24
 
 ### Changed